+++ RELEASENOTES.md
@@ -1,3 +1,13 @@
+## 1.5.1
+###### Updated Cmdlets:
+- **Import-LMNetscanTemplate**: 
+  -  Cmdlet now supports importing netscan templates for the following technologies:
+  -   Cisco Meraki
+  -   Cisco SDWAN
+  -   Cisco Catlayst Center Wireless
+  -   Juniper Mist
+  -   VMware vSphere
+
 ## 1.5
 ###### Updated Cmdlets:
 - **Import-LMNetscanTemplate**: